requirements.txt
Django==1.9.6
django-admin-bootstrapped==2.5.7
django-authtools==1.5.0
django-braces==1.8.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-environ==0.4.0
easy-thumbnails==2.3
gunicorn==19.5.0
Pillow==3.2.0
python-dateutil==2.5.3
python-magic==0.4.11
six==1.10.0
sqlparse==0.1.19
Werkzeug==0.11.9
